requirements.txt
asgiref
cffi
cryptography
Django
djangorestframework
pycparser
pytz
sqlparse
django-filter
drf_yasg
django-rest-authtoken
asgiref
gunicorn
psycopg2-binary
setuptools
wheel
djangorestframework-simplejwt